Fix bug when cancelling the file picker
Previously, when the file/folder picker was used, and then cancelled, the empty string would be returned. This would be passed to the pathfinder, which would interpret it as the current path, and return the source code's current working directory.

This behavior has been fixed, so that when the file picker is cancelled, there is no change to the explorer UI.